refactor(app): extract input handling + event dispatch into app_input.cpp
app.cpp was 5791 lines. Move the input/dispatch block (Input_OnMouse/Motion/Wheel/Special/Keyboard + up handlers, ApplyPropPositionDrag, InlineAutocomplete, DispatchEventBindings, ResetLevel/ResetScript) into app_input.cpp. Shared tuning constants (MOUSE_SENSITIVE, speeds, MK_* flags), the MOVEMENT_KEYS table, and small helpers (AiTextLineStarts/MaxChars, BuildRotMatZXY/ExtractEulerZXY) move to a private app_internal.h; the Win32 subclass globals and cursor/lookup helpers stay in core. app.cpp core drops to 3683 lines. Pure code motion: builds clean, runtime-verified on D:/IGI1 level 2 (558 objects, keybinds loaded, no errors).
